requirements.txt
python
streamlit
Pillow
torch
torchvision
opencv-python
numpy